modularize state in thread_local!
