+++ src/canisters-official-backend/src/core/mod.rs
@@ -1,0 +1,2 @@
+// src/core/mod.rs
+pub mod state;
