+++ src/canisters-official-backend/src/core/state/apikeys/types.rs
@@ -1,0 +1,12 @@
+
+// src/core/state/apikeys/types.rs
+
+use serde::{Deserialize, Serialize};
+
+
+#[derive(Debug, Clone, Serialize)]
+pub struct ApiKeyItem {
+    pub id: u32,
+    pub title: String,
+    pub completed: bool,
+}
