+++ src/canisters-official-backend/src/core/state/templates/mod.rs
@@ -1,0 +1,3 @@
+
+pub mod state;
+pub mod types;
