+++ src/canisters-official-backend/src/core/state/templates/types.rs
@@ -1,0 +1,9 @@
+// src/core/state/templates/types.rs
+use serde::{Serialize};
+
+#[derive(Debug, Clone, Serialize)]
+pub struct TemplateItem {
+    pub id: u32,
+    pub title: String,
+    pub completed: bool,
+}
